decompose did registry tests
